Advanced stage transitions, option to add stage transitions for the stage sets
