+++ red_task_handler/models/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
+from . import fs_stages
 from . import stage_sets
+from . import stage_transition
