+++ red_task_handler/security/ir.model.access.csv
@@ -10,3 +10,11 @@ access_selected_stage_mgr,selected_stages_mgr,red_task_handler.model_selected_st
 access_stages_emp,stages_emp,red_task_handler.model_fsm_stages,fieldservice.group_fsm_user,1,1,0,0
 access_stages_team_lead,stages_team_lead,red_task_handler.model_fsm_stages,fieldservice.group_fsm_dispatcher,1,1,1,0
 access_stages_mgr,access_mgr,red_task_handler.model_fsm_stages,fieldservice.group_fsm_manager,1,1,1,1
+
+access_stage_transition_person,stage_transition_person,red_task_handler.model_stage_transition_line,fieldservice.group_fsm_user,1,1,1,0
+access_stage_transition_team_lead,stage_transition_team_lead,red_task_handler.model_stage_transition_line,fieldservice.group_fsm_dispatcher,1,1,1,1
+access_stage_transition_mgr,stage_transition_mgr,red_task_handler.model_stage_transition_line,fieldservice.group_fsm_manager,1,1,1,1
+
+access_stage_trans_auth_person,stage_trans_auth_person,red_task_handler.model_stage_transition_authority,fieldservice.group_fsm_user,1,1,1,0
+access_stage_trans_auth_team_lead,stage_trans_auth_team_lead,red_task_handler.model_stage_transition_authority,fieldservice.group_fsm_dispatcher,1,1,1,1
+access_stage_trans_auth_mgr,stage_trans_auth_mgr,red_task_handler.model_stage_transition_authority,fieldservice.group_fsm_manager,1,1,1,1
