+++ red_task_handler/views/fs_stage_set.xml
@@ -1,43 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
     <data>
-        <!--fsm stages form view-->
-        <record id="fsm_stages_form" model="ir.ui.view">
-            <field name="name">StagesForm</field>
-            <field name="model">fsm.stages</field>
-            <field name="arch" type="xml">
-                <form>
-                    <sheet>
-                        <group>
-                            <group>
-                                <field name="name" />
-                                <field name="sequence" invisible="1" />
-                            </group>
-                            <group>
-                                <field name="fold" />
-                                <field name="ref_editable" invisible="1" />
-                                <field name="ref_name"
-                                       attrs="{'readonly': [('ref_editable','=',False)]}"
-                                />
-                            </group>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-        <!--fsm stages tree view-->
-        <record id="fsm_stages_tree" model="ir.ui.view">
-            <field name="name">StagesTree</field>
-            <field name="model">fsm.stages</field>
-            <field name="arch" type="xml">
-                <tree default_order="sequence asc">
-                    <field name="sequence" widget="handle" />
-                    <field name="name" />
-                    <field name="ref_name" />
-                </tree>
-            </field>
-        </record>
-
         <!--stage set form view-->
         <record id="fsm_stage_set_form" model="ir.ui.view">
             <field name="name">Stage Set Form</field>
@@ -95,6 +58,50 @@
                                                     self.state or parent_id.state : to access the stage of the corresponding records.<br />
                                                 </p>
                                             </div>
+                                            <!--Section that handles advanced transitions-->
+                                            <group>
+                                                <group string="Advanced Transitions">
+                                                    <field name="use_trasitions" />
+                                                </group>
+                                            </group>
+                                            <notebook>
+                                                <page string="Transitions"
+                                                      autofocus="1"
+                                                      attrs="{'invisible': [('use_trasitions', '=', False)]}">
+                                                    <field name="transition_ids"
+                                                           attrs="{'invisible': [('use_trasitions', '=', False)]}">
+                                                        <form string="Transition Details">
+                                                            <sheet>
+                                                                <group>
+                                                                    <group>
+                                                                        <field name="selected_stage_id"
+                                                                               invisible="1" />
+                                                                        <field name="dest_stage" />
+                                                                        <field name="transition_allowed"
+                                                                               widget="many2many_tags"
+                                                                        />
+                                                                    </group>
+                                                                    <group>
+                                                                        <field name="validate_transition" />
+                                                                    </group>
+                                                                </group>
+                                                            </sheet>
+                                                        </form>
+                                                        <tree editable="bottom">
+                                                            <field name="selected_stage_id"
+                                                                   invisible="1" />
+                                                            <field name="dest_stage"
+                                                                   required="1"
+                                                            />
+                                                            <field name="validate_transition" />
+                                                            <field name="transition_allowed"
+                                                                   widget="many2many_tags"
+                                                                   options="{'no_create': True}"
+                                                            />
+                                                        </tree>
+                                                    </field>
+                                                </page>
+                                            </notebook>
                                         </sheet>
                                     </form>
                                     <tree default_order="sequence asc">
@@ -126,20 +133,7 @@
                 </tree>
             </field>
         </record>
-        <!--stages action-->
-        <record id="action_fsm_stages" model="ir.actions.act_window">
-            <field name="name">Red FS Stages</field>
-            <field name="res_model">fsm.stages</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree, form</field>
-            <field name="view_ids"
-                   eval="[(5, 0, 0),
-                          (0, 0, {'view_mode': 'tree', 'view_id': ref('fsm_stages_tree')}),
-                          (0, 0, {'view_mode': 'form', 'view_id': ref('fsm_stages_form')})
-                          ]"
-            />
 
-        </record>
         <!--stage set action-->
         <record id="action_fsm_stage_sets" model="ir.actions.act_window">
             <field name="name">Stage Set</field>
@@ -152,13 +146,9 @@
                           (0, 0, {'view_mode': 'form', 'view_id': ref('fsm_stage_set_form')})
                           ]"
             />
-
         </record>
         <!--stage set menu-->
         <menuitem parent="fieldservice.config" id="fsm_stage_sets"
                   action="action_fsm_stage_sets" sequence="4" />
-        <!-- stages menu-->
-        <menuitem parent="fieldservice.config" id="fsm_stages"
-                  action="action_fsm_stages" sequence="5" />
     </data>
 </odoo>
